+++ Dealership/Dealership.Client/Commands/UserCommands/RemoveCarFromFavoritesCommand.cs
@@ -1,6 +1,7 @@
 ﻿using Dealership.Client.Commands.Abstract;
 using Dealership.Data.Models.Contracts;
 using Dealership.Services.Abstract;
+using System;
 
 namespace Dealership.Client.Commands.UserCommands
 {
@@ -15,6 +16,11 @@ namespace Dealership.Client.Commands.UserCommands
 
         public override string Execute(string[] parameters)
         {
+            if (parameters.Length != 1)
+            {
+                throw new ArgumentException("Invalid parameters");
+            }
+
             int carId = int.Parse(parameters[0]);
 
             var car = this.userService.RemoveCarFromFavorites(carId, this.UserSession.CurrentUser.Username);
